Reorganize code into separate files: VolumePage and TimeControl (Not Done)
The code was getting too long and messy with all of the Vuetify time control features.
Moving all code related to the Vuetify time control features into TimeControl.vue
